+++ src/main/kotlin/io/unthrottled/doki/settings/actors/LafAnimationActor.kt
@@ -3,6 +3,7 @@ package io.unthrottled.doki.settings.actors
 import com.intellij.openapi.util.registry.Registry
 import io.unthrottled.doki.config.ThemeConfig
 import io.unthrottled.doki.notification.UpdateNotification
+import io.unthrottled.doki.promotions.MessageBundle
 
 object LafAnimationActor {
   fun enableAnimation(enabled: Boolean) {
@@ -10,7 +11,10 @@ object LafAnimationActor {
       ThemeConfig.instance.isLafAnimation = enabled
       Registry.get("ide.intellij.laf.enable.animation").setValue(enabled)
       if (enabled) {
-        UpdateNotification.displayAnimationInstallMessage()
+        UpdateNotification.showDokiNotification(
+          MessageBundle.getMessage("notification.animation.install.title"),
+          MessageBundle.getMessage("notification.animation.install.body"),
+        )
       }
     }
   }