+++ src/main/kotlin/io/unthrottled/doki/settings/actors/ShowReadmeActor.kt
@@ -3,6 +3,7 @@ package io.unthrottled.doki.settings.actors
 import com.intellij.openapi.util.registry.Registry
 import io.unthrottled.doki.config.ThemeConfig
 import io.unthrottled.doki.notification.UpdateNotification
+import io.unthrottled.doki.promotions.MessageBundle
 
 object ShowReadmeActor {
   fun dontShowReadmeOnStartup(enabled: Boolean) {
@@ -10,7 +11,10 @@ object ShowReadmeActor {
       ThemeConfig.instance.isNotShowReadmeAtStartup = enabled
       Registry.get("ide.open.readme.md.on.startup").setValue(!enabled)
       if (enabled) {
-        UpdateNotification.displayReadmeInstallMessage()
+        UpdateNotification.showDokiNotification(
+          MessageBundle.message("notification.no.show.readme.title"),
+          MessageBundle.message("notification.no.show.readme.body"),
+        )
       }
     }
   }