+++ changelog/CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 ---
 
+# 74.2-1.0.5 [Update Notification Wallpaper Fix]
+
+- Fixed issue with multiple background images displaying over the update notification when the Doki Theme's wallpaper is activated ([#499](https://github.com/doki-theme/doki-theme-jetbrains/issues/499))
+
 # 74.2-1.0.4 [Small Enhancements]
 
 - Better Cucumber & Gherkin support